allow any model type as underlying model to interp model type
