+++ rust/routee-compass-powertrain/src/routee/prediction/interpolation/mod.rs
@@ -1,2 +1,3 @@
+pub mod bilinear_interp;
 pub mod interpolation_speed_grade_model;
 pub mod utils;
